requirements.txt
anyio==3.6.2
black==22.10.0
click==8.1.3
colorama==0.4.6
fastapi==0.87.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==0.4.3
# numpy==1.24.0
pandas==1.5.2
pathspec==0.10.2
platformdirs==2.5.4
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
python-dotenv==0.21.0
requests==2.27.1
beautifulsoup4==4.10.0
fastapi-redis-cache==0.2.5